api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
--index-url https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple

alembic==1.16.1
    # via flask-migrate
apscheduler==3.11.0
    # via
    #   -r requirements.in
    #   flask-apscheduler
blinker==1.9.0
    # via flask
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
colorama==0.4.6
    # via click
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-apscheduler
    #   flask-cors
    #   flask-migrate
    #   flask-sqlalchemy
    #   flask-wtf
flask-apscheduler==1.13.1
    # via -r requirements.in
flask-cors==6.0.0
    # via -r requirements.in
flask-migrate==4.1.0
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   -r requirements.in
    #   flask-migrate
flask-wtf==1.2.2
    # via -r requirements.in
greenlet==3.2.0
    # via sqlalchemy
idna==3.10
    # via requests
injector==0.22.0
    # via -r requirements.in
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.6
    # via flask
mako==1.3.10
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
marshmallow==4.0.0
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via flask-apscheduler
python-dotenv==1.1.0
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
rich==14.0.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.40
    # via
    #   alembic
    #   flask-sqlalchemy
typing-extensions==4.13.2
    # via
    #   alembic
    #   sqlalchemy
tzdata==2025.2
    # via tzlocal
tzlocal==5.3.1
    # via apscheduler
urllib3==2.4.0
    # via requests
waitress==3.0.2
    # via -r requirements.in
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
wtforms==3.2.1
    # via flask-wtf
